refactor(config): split llm settings
