+++ src/shared/infrastructure/config/settings_sections.py
@@ -5,14 +5,30 @@ from enum import StrEnum
 from importlib.metadata import PackageNotFoundError, version
 from ipaddress import IPv4Address
 from pathlib import Path
-from typing import Annotated, Any, Literal, cast
+from typing import Annotated, Any
 
 from pydantic import AliasChoices, Field, field_validator
-from pydantic_settings import BaseSettings, NoDecode, SettingsConfigDict
+from pydantic_settings import BaseSettings, NoDecode
+
+from src.shared.infrastructure.config.llm_settings import LLMSettings
+from src.shared.infrastructure.config.settings_base import (
+    LOCAL_DOTENV_FILE,
+    _settings_config,
+)
+
+__all__ = [
+    "APISettings",
+    "DEFAULT_SECRET_KEY",
+    "DatabaseSettings",
+    "Environment",
+    "LLMSettings",
+    "LOCAL_DOTENV_FILE",
+    "LogLevel",
+    "SecuritySettings",
+]
 
 _DEFAULT_SECRET_KEY_PARTS = ("change-me", "in-production", "32-char-long")
 DEFAULT_SECRET_KEY = "-".join(_DEFAULT_SECRET_KEY_PARTS)
-LOCAL_DOTENV_FILE = ".env.local"
 
 
 def _package_version() -> str:
@@ -24,29 +40,6 @@ def _package_version() -> str:
         return version("novel-engine")
     except PackageNotFoundError:
         return "0.3.1"
-
-
-def _settings_config(
-    *,
-    env_prefix: str,
-    env_nested_delimiter: str | None = None,
-    validate_default: bool | None = None,
-    populate_by_name: bool | None = None,
-) -> SettingsConfigDict:
-    config: dict[str, Any] = {
-        "env_file": LOCAL_DOTENV_FILE,
-        "env_file_encoding": "utf-8",
-        "env_prefix": env_prefix,
-        "extra": "ignore",
-        "case_sensitive": False,
-    }
-    if env_nested_delimiter is not None:
-        config["env_nested_delimiter"] = env_nested_delimiter
-    if validate_default is not None:
-        config["validate_default"] = validate_default
-    if populate_by_name is not None:
-        config["populate_by_name"] = populate_by_name
-    return cast(SettingsConfigDict, config)
 
 
 class Environment(StrEnum):
@@ -214,115 +207,3 @@ class SecuritySettings(BaseSettings):
         if isinstance(v, str):
             return [proxy.strip() for proxy in v.split(",") if proxy.strip()]
         return v if isinstance(v, list) else []
-
-
-class LLMSettings(BaseSettings):
-    model_config = _settings_config(env_prefix="LLM_")
-
-    provider: Literal["mock", "dashscope", "openai_compatible"] = Field(
-        default="mock",
-        description="LLM provider name",
-    )
-    model: str = Field(default="studio-copilot-v1", description="LLM model name")
-    api_key: str | None = Field(
-        default=None,
-        description="API key for OpenAI-compatible providers",
-        validation_alias=AliasChoices("LLM_API_KEY", "OPENAI_API_KEY"),
-    )
-    api_base: str | None = Field(
-        default=None,
-        description="Custom API base URL",
-        validation_alias=AliasChoices("LLM_API_BASE", "OPENAI_API_BASE"),
-    )
-    dashscope_api_key: str | None = Field(
-        default=None,
-        description="DashScope API key",
-        validation_alias=AliasChoices("DASHSCOPE_API_KEY"),
-    )
-    dashscope_api_base: str | None = Field(
-        default=None,
-        description="DashScope API base URL",
-        validation_alias=AliasChoices("DASHSCOPE_API_BASE"),
-    )
-    dashscope_model: str | None = Field(
-        default=None,
-        description="DashScope model override",
-        validation_alias=AliasChoices("DASHSCOPE_MODEL"),
-    )
-    dashscope_transport_mode: Literal[
-        "text_generation", "multimodal_generation", "responses"
-    ] = Field(
-        default="multimodal_generation",
-        description="DashScope native transport mode",
-        validation_alias=AliasChoices("DASHSCOPE_TRANSPORT_MODE"),
-    )
-    dashscope_review_model: str | None = Field(
-        default=None,
-        description="DashScope review model override",
-        validation_alias=AliasChoices("DASHSCOPE_REVIEW_MODEL"),
-    )
-    openai_compatible_model: str | None = Field(
-        default=None,
-        description="OpenAI-compatible model override",
-        validation_alias=AliasChoices("OPENAI_COMPATIBLE_MODEL"),
-    )
-    timeout: int = Field(
-        default=30, ge=5, le=300, description="Request timeout in seconds"
-    )
-    max_tokens: int = Field(
-        default=4096, ge=1, le=8192, description="Max tokens per request"
-    )
-    temperature: float = Field(
-        default=0.7, ge=0.0, le=2.0, description="Sampling temperature"
-    )
-    top_p: float = Field(default=0.95, ge=0.0, le=1.0, description="Top-p sampling")
-    top_k: int = Field(default=40, ge=1, le=100, description="Top-k sampling")
-    retry_attempts: int = Field(
-        default=3, ge=1, le=10, description="Number of retry attempts"
-    )
-    retry_delay: float = Field(
-        default=1.0, ge=0.1, le=10.0, description="Retry delay in seconds"
-    )
-
-    def resolved_api_key(
-        self,
-        provider_name: Literal["mock", "dashscope", "openai_compatible"] | str,
-    ) -> str | None:
-        if provider_name == "dashscope":
-            return self.dashscope_api_key
-        if provider_name == "openai_compatible":
-            return self.api_key
-        return None
-
-    def resolved_api_base(
-        self,
-        provider_name: Literal["mock", "dashscope", "openai_compatible"] | str,
-    ) -> str | None:
-        if provider_name == "dashscope":
-            return self.dashscope_api_base or "https://dashscope.aliyuncs.com/api/v1"
-        if provider_name == "openai_compatible":
-            return self.api_base or "https://api.openai.com/v1"
-        return None
-
-    def resolved_model(
-        self,
-        provider_name: Literal["mock", "dashscope", "openai_compatible"] | str,
-    ) -> str:
-        if provider_name == "dashscope":
-            return self.dashscope_model or self.model or "qwen3.5-flash"
-        if provider_name == "openai_compatible":
-            return self.openai_compatible_model or self.model or "gpt-4o-mini"
-        return self.model or "deterministic-story-v1"
-
-    def resolved_review_model(
-        self,
-        provider_name: Literal["mock", "dashscope", "openai_compatible"] | str,
-    ) -> str:
-        if provider_name == "dashscope":
-            return self.dashscope_review_model or self.resolved_model("dashscope")
-        return self.resolved_model(provider_name)
-
-    def resolved_dashscope_transport_mode(
-        self,
-    ) -> Literal["text_generation", "multimodal_generation", "responses"]:
-        return self.dashscope_transport_mode
